Build a basic production list view
Using a protocol `ProductionListViewable` to act as a view model implementation (a restriction with using protocols is that we can't cast a Core Data object's `title: String?` property into a `title: String` so we must use a different name).

Adding a `DemoProduction` object that conforms to the same protocol allows us to create a list of productions for SwiftUI/Xcode live previews without worrying about instantiating Core Data objects.